pyproject.toml
[tool.poetry]
name = "lanternfly-alerter"
version = "0.1.0"
description = "a service that alerts users of new spotted lanternfly observations on inaturalist"
authors = ["aholten"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
boto3 = "^1.28.61"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
pytest = "^7.4.2"
isort = "^5.12.0"
Flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
